requirements_det.txt
ansi2html==1.5.2
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==0.17.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==3.1.7
biopython==1.78
bleach==3.3.0
Brotli==1.0.9
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cssselect2==0.4.1
custom-inherit==2.3.0
cycler==0.10.0
dash==1.17.0
dash-bio==0.4.8
dash-bootstrap-components==0.10.7
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.20
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.0.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.6.3
jupyter-dash==0.3.1
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
multipledispatch==0.6.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
ncbi-genome-download==0.3.0
nest-asyncio==1.4.2
notebook==6.1.5
numpy==1.19.2
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.14.3
prometheus-client==0.8.0
prompt-toolkit==3.0.18
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
Pyphen==0.10.0
pypng==0.0.20
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==5.0.3
QtPy==1.9.0
reportlab==3.5.55
requests==2.24.0
retrying==1.3.3
rfc3986==1.4.0
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
Send2Trash==1.5.0
sip==6.0.3
six==1.15.0
sklearn==0.0
sniffio==1.2.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tinycss2==1.1.0
toml==0.10.2
tornado==6.0.4
toyplot==0.19.0
toytree==2.0.1
traitlets==5.0.5
urllib3==1.25.10
wcwidth==0.2.5
WeasyPrint==52.1
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xmltodict==0.12.0
zipp==3.4.1
